Add files via upload
The mass-spring-damper model consists of discrete mass nodes distributed throughout an object and interconnected via a network of springs and dampers. This model is well-suited for modelling object with complex material properties such as nonlinearity and viscoelasticity. 
Packages such as MATLAB may be used to run simulations of such models.
Objects may be described as volumetric meshes for simulation in this manner. As well as engineering simulation, these systems have applications in computer graphics and computer animation.